+++ src/main/java/kr/hhplus/be/commerce/application/message/publisher/MessagePublisher.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.commerce.domain.message.publisher;
+package kr.hhplus.be.commerce.application.message.publisher;
 
 import kr.hhplus.be.commerce.domain.message.enums.MessageType;
 import kr.hhplus.be.commerce.domain.message.model.message_payload.MessagePayload;
